httpapi: prevent against CSRF in logout
Tested using a simple HTML form like:

```
<html>
<body>
<form action="http://localhost:3080/.api/logout" method="post">
	<input type="submit" value="Submit">
</form>
</body>
</html>
```

Prior to this change, it would log you out. After this change, it shows an error:

```
cannot log out (no logged in user in context)
```

Additionally, normal logout was verified as working after this change.

We do not need additional CSRF prevention measures within httpapi, because we do not
allow any cookie-based authentication (see `httpapi/httpapi.go:29`). This specific CSRF
vulnerability was due to the fact that we would arbitrarily clear the user's session
regardless of whether or not they were authenticated (or not, in the case of a remote
site like the above HTML form).

Note that the lack of `handlerutil.UserMiddleware` additionally meant that:

- `serveSignup` could not detect logged in users, and thus would not prevent signed
  in users from creating new accounts (even though it explicitly tries to check for
  this).
- `serveForgotPassword` could not detect logged in users, and thus would allow
  requesting a forgot-password email while logged in.
- `servePasswordReset` could not detect logged in users, and thus would allow form
  submission of password change while logged in (this is the most scary, but I _think_
  because the token would have to be retrieved from the user's email account this could
  not be abused, I did not confirm this though).

+++ httpapi/auth.go
@@ -94,18 +94,27 @@ func serveSignup(w http.ResponseWriter, r *http.Request) error {
 }
 
 func serveLogout(w http.ResponseWriter, r *http.Request) error {
+	ctx := httpctx.FromRequest(r)
+	u := handlerutil.UserFromContext(ctx)
+
+	if u == nil || u.UID == 0 {
+		// SECURITY: If there is no authenticated user in the context, serve an
+		// error.
+		//
+		// This prevents CSRF attacks which allow external sites to log the
+		// user out by having them submit a form etc. Not a huge threat in the
+		// usual case, but it would still log users out and annoy them. Do not
+		// allow it.
+		return fmt.Errorf("cannot log out (no logged in user in context)")
+	}
+
+	// Delete their session.
 	appauth.DeleteSessionCookie(w)
 
-	ctx := httpctx.FromRequest(r)
-
-	currentUser := handlerutil.UserFromContext(ctx)
-	if currentUser != nil {
-		// If a user was logged in prior to navigating here, clear the user in
-		// the request context so that we don't show the logout page with the
-		// user's info.
-		ctx = handlerutil.ClearUser(ctx)
-		httpctx.SetForRequest(r, ctx)
-	}
+	// Clear the user in the request context so that we don't show the logout
+	// page with the user's info.
+	ctx = handlerutil.ClearUser(ctx)
+	httpctx.SetForRequest(r, ctx)
 
 	return writeJSON(w, &authResponse{Success: true})
 }
